backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
pymongo==4.6.0
python-jose[cryptography]==3.3.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
motor==3.3.2
httpx==0.25.2
collections-extended==2.0.2
